--- src/cht/test_util/tests.rs
@@ -1,861 +1,0 @@
-#[macro_export]
-macro_rules! write_test_cases_for_me {
-    ($m:ident) => {
-        #[test]
-        fn insertion() {
-            const MAX_VALUE: i32 = 512;
-
-            let map = $m::with_capacity(MAX_VALUE as usize);
-
-            for i in 0..MAX_VALUE {
-                assert_eq!(map.insert_entry_and(i, i, |_, v| v.clone()), None);
-
-                assert!(!map.is_empty());
-                assert_eq!(map.len(), (i + 1) as usize);
-
-                for j in 0..=i {
-                    assert_eq!(map.get(&j), Some(j));
-                    assert_eq!(map.insert_entry_and(j, j, |_, v| v.clone()), Some(j));
-                }
-
-                for k in i + 1..MAX_VALUE {
-                    assert_eq!(map.get(&k), None);
-                }
-            }
-
-            $crate::cht::test_util::run_deferred();
-        }
-
-        #[test]
-        fn growth() {
-            const MAX_VALUE: i32 = 512;
-
-            let map = $m::with_capacity(0);
-
-            for i in 0..MAX_VALUE {
-                assert_eq!(map.insert_entry_and(i, i, |_, v| v.clone()), None);
-
-                assert!(!map.is_empty());
-                assert_eq!(map.len(), (i + 1) as usize);
-
-                for j in 0..=i {
-                    assert_eq!(map.get(&j), Some(j));
-                    assert_eq!(map.insert_entry_and(j, j, |_, v| v.clone()), Some(j));
-                }
-
-                for k in i + 1..MAX_VALUE {
-                    assert_eq!(map.get(&k), None);
-                }
-            }
-
-            $crate::cht::test_util::run_deferred();
-        }
-
-        // Ignore this test and some other tests on 32-bit mips targets to avoid the following
-        // error on QEMU user space emulator:
-        //
-        //     memory allocation of 1052 bytes failed
-        //     process didn't exit successfully: ... (signal: 6, SIGABRT: process abort signal)
-        #[cfg_attr(mips, ignore)]
-        #[test]
-        fn concurrent_insertion() {
-            const MAX_VALUE: i32 = 512;
-            const NUM_THREADS: usize = 64;
-            const MAX_INSERTED_VALUE: i32 = (NUM_THREADS as i32) * MAX_VALUE;
-
-            let map = std::sync::Arc::new($m::with_capacity(MAX_INSERTED_VALUE as usize));
-            let barrier = std::sync::Arc::new(std::sync::Barrier::new(NUM_THREADS));
-
-            let threads: Vec<_> = (0..NUM_THREADS)
-                .map(|i| {
-                    let map = std::sync::Arc::clone(&map);
-                    let barrier = std::sync::Arc::clone(&barrier);
-
-                    std::thread::spawn(move || {
-                        barrier.wait();
-
-                        for j in (0..MAX_VALUE).map(|j| j + (i as i32 * MAX_VALUE)) {
-                            assert_eq!(map.insert_entry_and(j, j, |_, v| v.clone()), None);
-                        }
-                    })
-                })
-                .collect();
-
-            for result in threads.into_iter().map(std::thread::JoinHandle::join) {
-                assert!(result.is_ok());
-            }
-
-            assert!(!map.is_empty());
-            assert_eq!(map.len(), MAX_INSERTED_VALUE as usize);
-
-            for i in 0..MAX_INSERTED_VALUE {
-                assert_eq!(map.get(&i), Some(i));
-            }
-
-            $crate::cht::test_util::run_deferred();
-        }
-
-        #[cfg_attr(mips, ignore)]
-        #[test]
-        fn concurrent_growth() {
-            const MAX_VALUE: i32 = 512;
-            const NUM_THREADS: usize = 64;
-            const MAX_INSERTED_VALUE: i32 = (NUM_THREADS as i32) * MAX_VALUE;
-
-            let map = std::sync::Arc::new($m::with_capacity(0));
-            let barrier = std::sync::Arc::new(std::sync::Barrier::new(NUM_THREADS));
-
-            let threads: Vec<_> = (0..NUM_THREADS)
-                .map(|i| {
-                    let map = std::sync::Arc::clone(&map);
-                    let barrier = std::sync::Arc::clone(&barrier);
-
-                    std::thread::spawn(move || {
-                        barrier.wait();
-
-                        for j in (0..MAX_VALUE).map(|j| j + (i as i32 * MAX_VALUE)) {
-                            assert_eq!(map.insert_entry_and(j, j, |_, v| *v), None);
-                        }
-                    })
-                })
-                .collect();
-
-            for result in threads.into_iter().map(|t| t.join()) {
-                assert!(result.is_ok());
-            }
-
-            assert!(!map.is_empty());
-            assert_eq!(map.len(), MAX_INSERTED_VALUE as usize);
-
-            for i in 0..MAX_INSERTED_VALUE {
-                assert_eq!(map.get(&i), Some(i));
-            }
-
-            $crate::cht::test_util::run_deferred();
-        }
-
-        #[test]
-        fn removal() {
-            const MAX_VALUE: i32 = 512;
-
-            let map = $m::with_capacity(MAX_VALUE as usize);
-
-            for i in 0..MAX_VALUE {
-                assert_eq!(map.insert_entry_and(i, i, |_, v| *v), None);
-            }
-
-            for i in 0..MAX_VALUE {
-                assert_eq!(map.remove(&i), Some(i));
-            }
-
-            assert!(map.is_empty());
-            assert_eq!(map.len(), 0);
-
-            for i in 0..MAX_VALUE {
-                assert_eq!(map.get(&i), None);
-            }
-
-            $crate::cht::test_util::run_deferred();
-        }
-
-        #[cfg_attr(mips, ignore)]
-        #[test]
-        fn concurrent_removal() {
-            const MAX_VALUE: i32 = 512;
-            const NUM_THREADS: usize = 64;
-            const MAX_INSERTED_VALUE: i32 = (NUM_THREADS as i32) * MAX_VALUE;
-
-            let map = $m::with_capacity(MAX_INSERTED_VALUE as usize);
-
-            for i in 0..MAX_INSERTED_VALUE {
-                assert_eq!(map.insert_entry_and(i, i, |_, v| *v), None);
-            }
-
-            let map = std::sync::Arc::new(map);
-            let barrier = std::sync::Arc::new(std::sync::Barrier::new(NUM_THREADS));
-
-            let threads: Vec<_> = (0..NUM_THREADS)
-                .map(|i| {
-                    let map = std::sync::Arc::clone(&map);
-                    let barrier = std::sync::Arc::clone(&barrier);
-
-                    std::thread::spawn(move || {
-                        barrier.wait();
-
-                        for j in (0..MAX_VALUE).map(|j| j + (i as i32 * MAX_VALUE)) {
-                            assert_eq!(map.remove(&j), Some(j));
-                        }
-                    })
-                })
-                .collect();
-
-            for result in threads.into_iter().map(|t| t.join()) {
-                assert!(result.is_ok());
-            }
-
-            assert_eq!(map.len(), 0);
-
-            for i in 0..MAX_INSERTED_VALUE {
-                assert_eq!(map.get(&i), None);
-            }
-
-            $crate::cht::test_util::run_deferred();
-        }
-
-        #[cfg_attr(mips, ignore)]
-        #[test]
-        fn concurrent_insertion_and_removal() {
-            const MAX_VALUE: i32 = 512;
-            const NUM_THREADS: usize = 64;
-            const MAX_INSERTED_VALUE: i32 = (NUM_THREADS as i32) * MAX_VALUE * 2;
-            const INSERTED_MIDPOINT: i32 = MAX_INSERTED_VALUE / 2;
-
-            let map = $m::with_capacity(MAX_INSERTED_VALUE as usize);
-
-            for i in INSERTED_MIDPOINT..MAX_INSERTED_VALUE {
-                assert_eq!(map.insert_entry_and(i, i, |_, v| *v), None);
-            }
-
-            let map = std::sync::Arc::new(map);
-            let barrier = std::sync::Arc::new(std::sync::Barrier::new(NUM_THREADS * 2));
-
-            let insert_threads: Vec<_> = (0..NUM_THREADS)
-                .map(|i| {
-                    let map = std::sync::Arc::clone(&map);
-                    let barrier = std::sync::Arc::clone(&barrier);
-
-                    std::thread::spawn(move || {
-                        barrier.wait();
-
-                        for j in (0..MAX_VALUE).map(|j| j + (i as i32 * MAX_VALUE)) {
-                            assert_eq!(map.insert_entry_and(j, j, |_, v| *v), None);
-                        }
-                    })
-                })
-                .collect();
-
-            let remove_threads: Vec<_> = (0..NUM_THREADS)
-                .map(|i| {
-                    let map = std::sync::Arc::clone(&map);
-                    let barrier = std::sync::Arc::clone(&barrier);
-
-                    std::thread::spawn(move || {
-                        barrier.wait();
-
-                        for j in
-                            (0..MAX_VALUE).map(|j| INSERTED_MIDPOINT + j + (i as i32 * MAX_VALUE))
-                        {
-                            assert_eq!(map.remove(&j), Some(j));
-                        }
-                    })
-                })
-                .collect();
-
-            for result in insert_threads
-                .into_iter()
-                .chain(remove_threads.into_iter())
-                .map(|t| t.join())
-            {
-                assert!(result.is_ok());
-            }
-
-            assert!(!map.is_empty());
-            assert_eq!(map.len(), INSERTED_MIDPOINT as usize);
-
-            for i in 0..INSERTED_MIDPOINT {
-                assert_eq!(map.get(&i), Some(i));
-            }
-
-            for i in INSERTED_MIDPOINT..MAX_INSERTED_VALUE {
-                assert_eq!(map.get(&i), None);
-            }
-
-            $crate::cht::test_util::run_deferred();
-        }
-
-        #[cfg_attr(mips, ignore)]
-        #[test]
-        fn concurrent_growth_and_removal() {
-            const MAX_VALUE: i32 = 512;
-            const NUM_THREADS: usize = 64;
-            const MAX_INSERTED_VALUE: i32 = (NUM_THREADS as i32) * MAX_VALUE * 2;
-            const INSERTED_MIDPOINT: i32 = MAX_INSERTED_VALUE / 2;
-
-            let map = $m::with_capacity(INSERTED_MIDPOINT as usize);
-
-            for i in INSERTED_MIDPOINT..MAX_INSERTED_VALUE {
-                assert_eq!(map.insert_entry_and(i, i, |_, v| *v), None);
-            }
-
-            let map = std::sync::Arc::new(map);
-            let barrier = std::sync::Arc::new(std::sync::Barrier::new(NUM_THREADS * 2));
-
-            let insert_threads: Vec<_> = (0..NUM_THREADS)
-                .map(|i| {
-                    let map = std::sync::Arc::clone(&map);
-                    let barrier = std::sync::Arc::clone(&barrier);
-
-                    std::thread::spawn(move || {
-                        barrier.wait();
-
-                        for j in (0..MAX_VALUE).map(|j| j + (i as i32 * MAX_VALUE)) {
-                            assert_eq!(map.insert_entry_and(j, j, |_, v| *v), None);
-                        }
-                    })
-                })
-                .collect();
-
-            let remove_threads: Vec<_> = (0..NUM_THREADS)
-                .map(|i| {
-                    let map = std::sync::Arc::clone(&map);
-                    let barrier = std::sync::Arc::clone(&barrier);
-
-                    std::thread::spawn(move || {
-                        barrier.wait();
-
-                        for j in
-                            (0..MAX_VALUE).map(|j| INSERTED_MIDPOINT + j + (i as i32 * MAX_VALUE))
-                        {
-                            assert_eq!(map.remove(&j), Some(j));
-                        }
-                    })
-                })
-                .collect();
-
-            for result in insert_threads
-                .into_iter()
-                .chain(remove_threads.into_iter())
-                .map(std::thread::JoinHandle::join)
-            {
-                assert!(result.is_ok());
-            }
-
-            assert!(!map.is_empty());
-            assert_eq!(map.len(), INSERTED_MIDPOINT as usize);
-
-            for i in 0..INSERTED_MIDPOINT {
-                assert_eq!(map.get(&i), Some(i));
-            }
-
-            for i in INSERTED_MIDPOINT..MAX_INSERTED_VALUE {
-                assert_eq!(map.get(&i), None);
-            }
-
-            $crate::cht::test_util::run_deferred();
-        }
-
-        #[test]
-        fn insert_with_or_modify() {
-            let map = $m::with_capacity(0);
-
-            assert_eq!(map.insert_with_or_modify("foo", || 1, |_, x| x + 1), None);
-            assert_eq!(map.get("foo"), Some(1));
-
-            assert_eq!(
-                map.insert_with_or_modify("foo", || 1, |_, x| x + 1),
-                Some(1)
-            );
-            assert_eq!(map.get("foo"), Some(2));
-
-            $crate::cht::test_util::run_deferred();
-        }
-
-        #[cfg_attr(mips, ignore)]
-        #[test]
-        fn concurrent_insert_with_or_modify() {
-            const NUM_THREADS: usize = 64;
-            const MAX_VALUE: i32 = 512;
-
-            let map = std::sync::Arc::new($m::with_capacity(0));
-            let barrier = std::sync::Arc::new(std::sync::Barrier::new(NUM_THREADS));
-
-            let threads: Vec<_> = (0..NUM_THREADS)
-                .map(|_| {
-                    let map = std::sync::Arc::clone(&map);
-                    let barrier = std::sync::Arc::clone(&barrier);
-
-                    std::thread::spawn(move || {
-                        barrier.wait();
-
-                        for j in 0..MAX_VALUE {
-                            map.insert_with_or_modify(j, || 1, |_, x| x + 1);
-                        }
-                    })
-                })
-                .collect();
-
-            for result in threads.into_iter().map(std::thread::JoinHandle::join) {
-                assert!(result.is_ok());
-            }
-
-            assert_eq!(map.len(), MAX_VALUE as usize);
-
-            for i in 0..MAX_VALUE {
-                assert_eq!(map.get(&i), Some(NUM_THREADS as i32));
-            }
-
-            $crate::cht::test_util::run_deferred();
-        }
-
-        #[cfg_attr(mips, ignore)]
-        #[test]
-        fn concurrent_overlapped_insertion() {
-            const NUM_THREADS: usize = 64;
-            const MAX_VALUE: i32 = 512;
-
-            let map = std::sync::Arc::new($m::with_capacity(MAX_VALUE as usize));
-            let barrier = std::sync::Arc::new(std::sync::Barrier::new(NUM_THREADS));
-
-            let threads: Vec<_> = (0..NUM_THREADS)
-                .map(|_| {
-                    let map = std::sync::Arc::clone(&map);
-                    let barrier = std::sync::Arc::clone(&barrier);
-
-                    std::thread::spawn(move || {
-                        barrier.wait();
-
-                        for j in 0..MAX_VALUE {
-                            map.insert_entry_and(j, j, |_, v| *v);
-                        }
-                    })
-                })
-                .collect();
-
-            for result in threads.into_iter().map(std::thread::JoinHandle::join) {
-                assert!(result.is_ok());
-            }
-
-            assert_eq!(map.len(), MAX_VALUE as usize);
-
-            for i in 0..MAX_VALUE {
-                assert_eq!(map.get(&i), Some(i));
-            }
-
-            $crate::cht::test_util::run_deferred();
-        }
-
-        // Ignore this test on 32-bit mips and armv5te targets to avoid the following
-        // error on QEMU user space emulator:
-        //
-        // (mips)
-        //     memory allocation of 1052 bytes failed
-        //     process didn't exit successfully: ... (signal: 6, SIGABRT: process abort signal)
-        //
-        // (armv5te)
-        //     process didn't exit successfully: ... (signal: 4, SIGILL: illegal instruction)
-        //
-        #[cfg_attr(any(armv5te, mips), ignore)]
-        #[test]
-        fn concurrent_overlapped_growth() {
-            const NUM_THREADS: usize = 64;
-            const MAX_VALUE: i32 = 512;
-
-            let map = std::sync::Arc::new($m::with_capacity(1));
-            let barrier = std::sync::Arc::new(std::sync::Barrier::new(NUM_THREADS));
-
-            let threads: Vec<_> = (0..NUM_THREADS)
-                .map(|_| {
-                    let map = std::sync::Arc::clone(&map);
-                    let barrier = std::sync::Arc::clone(&barrier);
-
-                    std::thread::spawn(move || {
-                        barrier.wait();
-
-                        for j in 0..MAX_VALUE {
-                            map.insert_entry_and(j, j, |_, v| *v);
-                        }
-                    })
-                })
-                .collect();
-
-            for result in threads.into_iter().map(std::thread::JoinHandle::join) {
-                assert!(result.is_ok());
-            }
-
-            assert_eq!(map.len(), MAX_VALUE as usize);
-
-            for i in 0..MAX_VALUE {
-                assert_eq!(map.get(&i), Some(i));
-            }
-
-            $crate::cht::test_util::run_deferred();
-        }
-
-        #[cfg_attr(mips, ignore)]
-        #[test]
-        fn concurrent_overlapped_removal() {
-            const NUM_THREADS: usize = 64;
-            const MAX_VALUE: i32 = 512;
-
-            let map = $m::with_capacity(MAX_VALUE as usize);
-
-            for i in 0..MAX_VALUE {
-                map.insert_entry_and(i, i, |_, v| *v);
-            }
-
-            let map = std::sync::Arc::new(map);
-            let barrier = std::sync::Arc::new(std::sync::Barrier::new(NUM_THREADS));
-
-            let threads: Vec<_> = (0..NUM_THREADS)
-                .map(|_| {
-                    let map = std::sync::Arc::clone(&map);
-                    let barrier = std::sync::Arc::clone(&barrier);
-
-                    std::thread::spawn(move || {
-                        barrier.wait();
-
-                        for j in 0..MAX_VALUE {
-                            let prev_value = map.remove(&j);
-
-                            if let Some(v) = prev_value {
-                                assert_eq!(v, j);
-                            }
-                        }
-                    })
-                })
-                .collect();
-
-            for result in threads.into_iter().map(std::thread::JoinHandle::join) {
-                assert!(result.is_ok());
-            }
-
-            assert!(map.is_empty());
-            assert_eq!(map.len(), 0);
-
-            for i in 0..MAX_VALUE {
-                assert_eq!(map.get(&i), None);
-            }
-
-            $crate::cht::test_util::run_deferred();
-        }
-
-        #[test]
-        fn drop_value() {
-            let key_parent = std::sync::Arc::new($crate::cht::test_util::DropNotifier::new());
-            let value_parent = std::sync::Arc::new($crate::cht::test_util::DropNotifier::new());
-
-            {
-                let map = $m::with_capacity(0);
-
-                assert_eq!(
-                    map.insert_entry_and(
-                        $crate::cht::test_util::NoisyDropper::new(
-                            std::sync::Arc::clone(&key_parent),
-                            0
-                        ),
-                        $crate::cht::test_util::NoisyDropper::new(
-                            std::sync::Arc::clone(&value_parent),
-                            0
-                        ),
-                        |_, _| ()
-                    ),
-                    None
-                );
-                assert!(!map.is_empty());
-                assert_eq!(map.len(), 1);
-                map.get_key_value_and(&0, |_k, v| assert_eq!(v, &0));
-
-                map.remove_entry_if_and(&0, |_, _| true, |_k, v| assert_eq!(v, &0));
-                assert!(map.is_empty());
-                assert_eq!(map.len(), 0);
-                assert_eq!(map.get_key_value_and(&0, |_, _| ()), None);
-
-                $crate::cht::test_util::run_deferred();
-
-                assert!(!key_parent.was_dropped());
-                assert!(value_parent.was_dropped());
-            }
-
-            $crate::cht::test_util::run_deferred();
-
-            assert!(key_parent.was_dropped());
-            assert!(value_parent.was_dropped());
-        }
-
-        #[test]
-        fn drop_many_values() {
-            const NUM_VALUES: usize = 1 << 16;
-
-            let key_parents: Vec<_> = std::iter::repeat_with(|| {
-                std::sync::Arc::new($crate::cht::test_util::DropNotifier::new())
-            })
-            .take(NUM_VALUES)
-            .collect();
-            let value_parents: Vec<_> = std::iter::repeat_with(|| {
-                std::sync::Arc::new($crate::cht::test_util::DropNotifier::new())
-            })
-            .take(NUM_VALUES)
-            .collect();
-
-            {
-                let map = $m::with_capacity(0);
-                assert!(map.is_empty());
-                assert_eq!(map.len(), 0);
-
-                for (i, (this_key_parent, this_value_parent)) in
-                    key_parents.iter().zip(value_parents.iter()).enumerate()
-                {
-                    assert_eq!(
-                        map.insert_entry_and(
-                            $crate::cht::test_util::NoisyDropper::new(
-                                std::sync::Arc::clone(&this_key_parent),
-                                i
-                            ),
-                            $crate::cht::test_util::NoisyDropper::new(
-                                std::sync::Arc::clone(&this_value_parent),
-                                i
-                            ),
-                            |_, _| ()
-                        ),
-                        None
-                    );
-
-                    assert!(!map.is_empty());
-                    assert_eq!(map.len(), i + 1);
-                }
-
-                for i in 0..NUM_VALUES {
-                    assert_eq!(
-                        map.get_key_value_and(&i, |k, v| {
-                            assert_eq!(*k, i);
-                            assert_eq!(*v, i);
-                        }),
-                        Some(())
-                    );
-                }
-
-                for i in 0..NUM_VALUES {
-                    assert_eq!(
-                        map.remove_entry_if_and(
-                            &i,
-                            |_, _| true,
-                            |k, v| {
-                                assert_eq!(*k, i);
-                                assert_eq!(*v, i);
-                            }
-                        ),
-                        Some(())
-                    );
-                }
-
-                assert!(map.is_empty());
-                assert_eq!(map.len(), 0);
-
-                $crate::cht::test_util::run_deferred();
-
-                let live_key_count =
-                    NUM_VALUES - key_parents.iter().filter(|k| k.was_dropped()).count();
-                let bucket_array_len = map.capacity() * 2;
-                assert_eq!(bucket_array_len, map.num_segments() * 128 * 2);
-                assert!(live_key_count <= bucket_array_len / 10);
-
-                for this_value_parent in value_parents.iter() {
-                    assert!(this_value_parent.was_dropped());
-                }
-
-                for i in 0..NUM_VALUES {
-                    assert_eq!(map.get_key_value_and(&i, |_, _| ()), None);
-                }
-            }
-
-            $crate::cht::test_util::run_deferred();
-
-            for this_key_parent in key_parents.into_iter() {
-                assert!(this_key_parent.was_dropped());
-            }
-
-            for this_value_parent in value_parents.into_iter() {
-                assert!(this_value_parent.was_dropped());
-            }
-        }
-
-        #[test]
-        fn drop_many_values_concurrent() {
-            const NUM_THREADS: usize = 64;
-            const NUM_VALUES_PER_THREAD: usize = 512;
-            const NUM_VALUES: usize = NUM_THREADS * NUM_VALUES_PER_THREAD;
-
-            let key_parents: std::sync::Arc<Vec<_>> = std::sync::Arc::new(
-                std::iter::repeat_with(|| {
-                    std::sync::Arc::new($crate::cht::test_util::DropNotifier::new())
-                })
-                .take(NUM_VALUES)
-                .collect(),
-            );
-            let value_parents: std::sync::Arc<Vec<_>> = std::sync::Arc::new(
-                std::iter::repeat_with(|| {
-                    std::sync::Arc::new($crate::cht::test_util::DropNotifier::new())
-                })
-                .take(NUM_VALUES)
-                .collect(),
-            );
-
-            {
-                let map = std::sync::Arc::new($m::with_capacity(0));
-                assert!(map.is_empty());
-                assert_eq!(map.len(), 0);
-
-                let barrier = std::sync::Arc::new(std::sync::Barrier::new(NUM_THREADS));
-
-                let handles: Vec<_> = (0..NUM_THREADS)
-                    .map(|i| {
-                        let map = std::sync::Arc::clone(&map);
-                        let barrier = std::sync::Arc::clone(&barrier);
-                        let key_parents = std::sync::Arc::clone(&key_parents);
-                        let value_parents = std::sync::Arc::clone(&value_parents);
-
-                        std::thread::spawn(move || {
-                            barrier.wait();
-
-                            let these_key_parents = &key_parents
-                                [i * NUM_VALUES_PER_THREAD..(i + 1) * NUM_VALUES_PER_THREAD];
-                            let these_value_parents = &value_parents
-                                [i * NUM_VALUES_PER_THREAD..(i + 1) * NUM_VALUES_PER_THREAD];
-
-                            for (j, (this_key_parent, this_value_parent)) in these_key_parents
-                                .iter()
-                                .zip(these_value_parents.iter())
-                                .enumerate()
-                            {
-                                let key_value = i * NUM_VALUES_PER_THREAD + j;
-
-                                assert_eq!(
-                                    map.insert_entry_and(
-                                        $crate::cht::test_util::NoisyDropper::new(
-                                            std::sync::Arc::clone(&this_key_parent),
-                                            key_value as i32
-                                        ),
-                                        $crate::cht::test_util::NoisyDropper::new(
-                                            std::sync::Arc::clone(&this_value_parent),
-                                            key_value as i32
-                                        ),
-                                        |_, _| ()
-                                    ),
-                                    None
-                                );
-                            }
-                        })
-                    })
-                    .collect();
-
-                for result in handles.into_iter().map(std::thread::JoinHandle::join) {
-                    assert!(result.is_ok());
-                }
-
-                assert!(!map.is_empty());
-                assert_eq!(map.len(), NUM_VALUES);
-
-                $crate::cht::test_util::run_deferred();
-
-                for this_key_parent in key_parents.iter() {
-                    assert!(!this_key_parent.was_dropped());
-                }
-
-                for this_value_parent in value_parents.iter() {
-                    assert!(!this_value_parent.was_dropped());
-                }
-
-                for i in (0..NUM_VALUES).map(|i| i as i32) {
-                    assert_eq!(
-                        map.get_key_value_and(&i, |k, v| {
-                            assert_eq!(*k, i);
-                            assert_eq!(*v, i);
-                        }),
-                        Some(())
-                    );
-                }
-
-                let handles: Vec<_> = (0..NUM_THREADS)
-                    .map(|i| {
-                        let map = std::sync::Arc::clone(&map);
-                        let barrier = std::sync::Arc::clone(&barrier);
-
-                        std::thread::spawn(move || {
-                            barrier.wait();
-
-                            for j in 0..NUM_VALUES_PER_THREAD {
-                                let key_value = (i * NUM_VALUES_PER_THREAD + j) as i32;
-
-                                assert_eq!(
-                                    map.remove_entry_if_and(
-                                        &key_value,
-                                        |_, _| true,
-                                        |k, v| {
-                                            assert_eq!(*k, key_value);
-                                            assert_eq!(*v, key_value);
-                                        }
-                                    ),
-                                    Some(())
-                                );
-                            }
-                        })
-                    })
-                    .collect();
-
-                for result in handles.into_iter().map(std::thread::JoinHandle::join) {
-                    assert!(result.is_ok());
-                }
-
-                assert!(map.is_empty());
-                assert_eq!(map.len(), 0);
-
-                $crate::cht::test_util::run_deferred();
-
-                let live_key_count =
-                    NUM_VALUES - key_parents.iter().filter(|k| k.was_dropped()).count();
-                let bucket_array_len = map.capacity() * 2;
-                assert_eq!(bucket_array_len, map.num_segments() * 128 * 2);
-                assert!(live_key_count <= bucket_array_len / 10);
-
-                for this_value_parent in value_parents.iter() {
-                    assert!(this_value_parent.was_dropped());
-                }
-
-                for i in (0..NUM_VALUES).map(|i| i as i32) {
-                    assert_eq!(map.get_key_value_and(&i, |_, _| ()), None);
-                }
-            }
-
-            $crate::cht::test_util::run_deferred();
-
-            for this_key_parent in key_parents.iter() {
-                assert!(this_key_parent.was_dropped());
-            }
-
-            for this_value_parent in value_parents.iter() {
-                assert!(this_value_parent.was_dropped());
-            }
-        }
-
-        #[test]
-        fn remove_if() {
-            const NUM_VALUES: i32 = 512;
-
-            let is_even = |_: &i32, v: &i32| *v % 2 == 0;
-
-            let map = $m::with_capacity(0);
-
-            for i in 0..NUM_VALUES {
-                assert_eq!(map.insert_entry_and(i, i, |_, v| *v), None);
-            }
-
-            for i in 0..NUM_VALUES {
-                if is_even(&i, &i) {
-                    assert_eq!(map.remove_if(&i, is_even), Some(i));
-                } else {
-                    assert_eq!(map.remove_if(&i, is_even), None);
-                }
-            }
-
-            for i in (0..NUM_VALUES).filter(|i| i % 2 == 0) {
-                assert_eq!(map.get(&i), None);
-            }
-
-            for i in (0..NUM_VALUES).filter(|i| i % 2 != 0) {
-                assert_eq!(map.get(&i), Some(i));
-            }
-
-            $crate::cht::test_util::run_deferred();
-        }
-    };
-}
